Prefix cheerio-argument-having functions with 'extract' for consistency and to avoid naming conflicts
