+++ src/lib/extractBaseHref.ts
@@ -3,7 +3,7 @@ import { buildAbsoluteURL } from 'url-toolkit';
 
 import isAbsoluteUrl from './isAbsoluteUrl';
 
-const getBaseHref = ($: CheerioAPI, documentHref: string): string => {
+const extractBaseHref = ($: CheerioAPI, documentHref: string): string => {
   const [$base] = Array.from($('base'));
   const baseHref = $base && $($base).attr('href');
   const isBaseHrefAbsolute = baseHref ? isAbsoluteUrl(baseHref) : false;
@@ -15,4 +15,4 @@ const getBaseHref = ($: CheerioAPI, documentHref: string): string => {
     : documentHref;
 };
 
-export default getBaseHref;
+export default extractBaseHref;
