+++ src/lib/extractFavicons.ts
@@ -3,19 +3,19 @@ import { buildAbsoluteURL } from 'url-toolkit';
 
 import { Favicon } from '../types';
 
-import getBaseHref from './getBaseHref';
+import extractBaseHref from './extractBaseHref';
 import isFaviconUrl from './isFaviconUrl';
 import parseLinkSizes from './parseLinkSizes';
 import sortFavicons from './sortFavicons';
 
-const getFavicons = ($: CheerioAPI, documentHref: string): Favicon[] => {
+const extractFavicons = ($: CheerioAPI, documentHref: string): Favicon[] => {
   const $links = [
     ...$('link[rel="shortcut icon"]'),
     ...$('link[rel="icon"]'),
     ...$('link[rel="apple-touch-icon"]'),
     ...$('link[rel="image_src"]'), // stackoverflow.com uses this attribute
   ];
-  const baseHref = getBaseHref($, documentHref);
+  const baseHref = extractBaseHref($, documentHref);
   const faviconLinks: Favicon[] = $links
     .filter(($link) => isFaviconUrl($($link).attr('href'), documentHref))
     .map(($link) => {
@@ -30,4 +30,4 @@ const getFavicons = ($: CheerioAPI, documentHref: string): Favicon[] => {
   return sortFavicons(faviconLinks);
 };
 
-export default getFavicons;
+export default extractFavicons;
